change namespaces of auto moderation actions, rules and processors
