+++ lib/decidim/decidim_awesome.rb
@@ -5,7 +5,6 @@ require "decidim/decidim_awesome/admin"
 require "decidim/decidim_awesome/engine"
 require "decidim/decidim_awesome/admin_engine"
 require "decidim/decidim_awesome/context_analyzers"
-require "decidim/decidim_awesome/processors/moderation_processor"
 require "decidim/decidim_awesome/map_component/engine"
 require "decidim/decidim_awesome/map_component/admin_engine"
 require "decidim/decidim_awesome/map_component/component"
