--- lib/decidim/decidim_awesome/moderation_actions.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module DecidimAwesome
-    module ModerationActions
-      autoload :ModerateAndHide, "decidim/decidim_awesome/moderation_actions/moderate_and_hide"
-    end
-  end
-end
