--- lib/decidim/decidim_awesome/moderation_rules.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module DecidimAwesome
-    module ModerationRules
-      autoload :WordFilter, "decidim/decidim_awesome/moderation_rules/word_filter"
-    end
-  end
-end
