--- lib/decidim/decidim_awesome/processors.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Decidim
-  module DecidimAwesome
-    module Processors
-      autoload :ModerationProcessor, "decidim/decidim_awesome/processors/moderation_processor"
-    end
-  end
-end
